Create FIBONACCI_MATHS.txt
The Mathematical Magic of the Fibonacci Numbers
This page looks at some patterns in the Fibonacci numbers themselves, from the digits in the numbers to their factors and multiples and which are prime numbers. There is an unexpected pattern in the initial digits too. We also relate Fibonacci numbers to Pascal's triangle via the original rabbit problem that Fibonacci used to introduce the series we now call by his name. We can also make the Fibonacci numbers appear in a decimal fraction, introduce you to an easily learned number magic trick that only works with Fibonacci-like series numbers, see how Pythagoras' Theorem and right-angled triangles such as 3-4-5 have connections with the Fibonacci numbers and then give you lots of hints and suggestions for finding more number patterns of your own.